% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_function.R
\name{length_conv}
\alias{length_conv}
\title{Length Function}
\usage{
length_conv(n, from, to)
}
\arguments{
\item{n}{is the value of the initial lengeth you want to convert}

\item{from}{the units of the initial values, inch= inches, cm = centemeters, m = meter, km = kilometer, mm = milimeter, mile = mile, yd = yards, ft = foot, microm = micrometer, nm = nanometer}

\item{to}{the units you want to convert to}
}
\value{
the value of the conversion in the units of "to"
}
\description{
Length Function
}
\examples{
length_conv(5,"yd","mile"),
length_conv(23, "mile", "nm")
}
