% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test func}
\usage{
test(n)
}
\arguments{
\item{n}{#text}
}
\value{
#int
}
\description{
Function that converts numerical values from and to seconds, minutes, hours, days, weeks, months, years.
}
\examples{
timeConvert(5,minutes, hours) =  this is a test for the timeConvert function
}
