% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight_conv}
\alias{weight_conv}
\title{Convert from one unit of weight measurement to another}
\usage{
weight_conv(n, from, to)
}
\arguments{
\item{n}{#quantity to change}

\item{from}{# unit of quantity}

\item{to}{#desired unit}
}
\value{
#returns desired quantity in the desired unit
}
\description{
Convert from one unit of weight measurement to another
}
\examples{
weight_conv(500,"lb","kg") = 226.796
weight_conv(10,"kg","T") = 0.01102312
weight_conv(10,"g","mg") = 10000
}
