% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeConvert.R
\name{timeConvert}
\alias{timeConvert}
\title{timeConvert func}
\usage{
timeConvert(n, from, to)
}
\arguments{
\item{n}{#quantity to change}

\item{from}{#unit quantity}

\item{to}{#unit desired}
}
\value{
#new quantity in right units
}
\description{
timeConvert func
}
\examples{
test(4)=  this is a test if 4 works
}
